chore: address review feedback
- Fix announceChecked not syncing with Firestore updates
- Extract duplicate statusToView/routeToView to lib/routing.ts
- Extract duplicate isCompleteGroup to store/types.ts
- Move initPools from store/types.ts to lib/roles.ts
- Compose resetSession from sub-reset methods
- Remove unused useDiscord hook
- Remove unused onNavigate prop from ResultsView
- Use individual Zustand selectors in useIdentity
- Add comment explaining hidden #next-btn for Playwright
- Fix stale closure: checkForPendingReveals in callback deps
- Add navigateTo to useEffect deps in App.tsx
