+++ activity/src/store/types.ts
@@ -1,4 +1,8 @@
-import { WoWGroup, WoWPlayer, WheelEntry, GuildData, ChannelData } from '../types';
+import { WoWGroup, WheelEntry, GuildData, ChannelData } from '../types';
+
+export function isCompleteGroup(group: WoWGroup): boolean {
+  return group.tank !== null && group.healer !== null && group.dps.length === 3;
+}
 
 export type ViewName = 'home' | 'channels' | 'lobby' | 'wheels' | 'results';
 
@@ -73,20 +77,4 @@ export interface GroupCardData {
   index: number;
   label?: string;
   hideEmpty?: boolean;
-}
-
-export function initPools(players: WoWPlayer[]): { tanks: WheelEntry[]; healers: WheelEntry[]; dps: WheelEntry[] } {
-  const tanks = players
-    .filter((p) => p.roles.tankMain || p.roles.offtank)
-    .map((p) => ({ name: p.name, isOffspec: !p.roles.tankMain }));
-
-  const healers = players
-    .filter((p) => p.roles.healerMain || p.roles.offhealer)
-    .map((p) => ({ name: p.name, isOffspec: !p.roles.healerMain }));
-
-  const dps = players
-    .filter((p) => p.roles.dpsMain || p.roles.offdps)
-    .map((p) => ({ name: p.name, isOffspec: !p.roles.dpsMain }));
-
-  return { tanks, healers, dps };
 }
